moved `HeaderSearchResultDropdown` to `src/containers`
