+++ src/hooks/useModal.tsx
@@ -72,7 +72,7 @@ export function ModalContainer() {
 
 /**
  * the API for showing the modal, only thing needed is the showModal function that accepts
- * content as the first parameter, that'HeaderSearchResultsDropdown all it takes to render a modal
+ * content as the first parameter, that's all it takes to render a modal
  */
 export default function useModal() {
   const { show, hide, isOpen } = React.useContext(ModalContext);
